Converted Damage/Health System, Balanced Weapons
Health/Damage system now uses Float values instead of Int values.
Made the Remington 870 and AA12 stronger while lowering the strength of the Barrett.

+++ TopDownHordeGame_UnityProject/Assets/Scripts/Weapon/Grenade.cs
@@ -6,7 +6,7 @@ public class Grenade : MonoBehaviour
 {
     public GameObject explosionPrefab;
     private GameObject explosionObj;
-    private int balanceDamage;
+    private float balanceDamage;
     private float throwStrength;
     private float flySpeed;
     private Vector3 rotationTemp;
@@ -16,7 +16,7 @@ public class Grenade : MonoBehaviour
 
     private GameObject owner;
 
-    public void Init(GameObject newOwner, Vector2 movementDir, int damage, float radius, float speed, float knockback, float explodeTime)
+    public void Init(GameObject newOwner, Vector2 movementDir, float damage, float radius, float speed, float knockback, float explodeTime)
     {
         owner = newOwner; 
         transform.position = newOwner.transform.position;
